- 2db207ef16d548380e4f5cd1d583546d0b5f7d61 Run clang-format on all outbound Copybara-transformed code. by Daniel Katz <katzdm@google.com>   - f9f24ecdf159c232b68a72b3be4d66854489f13e Run clang-format on all outbound Copybara-transformed code. by Daniel Katz <katzdm@google.com>   - 54b1d34753a2b6999bb65e0c05255bde7de8cfb5 Properly indent the PROPERTY FOLDER calls in AbseilHelper... by Jon Cohen <cohenjon@google.com>   - 540ce6bab1a7c81a15d7244b9effcb1c1378911b Fix -Wgnu-zero-variadic-macro-arguments warning under cla... by Derek Mauro <dmauro@google.com>   - cc82ee3f18e882bfcba911ebc1b8ea39983f319b Suppress MSVC unreachable code warning in optional.h by Abseil Team <absl-team@google.com>   - c1176bed2a7a2d20e72db2601f1044c097d8078c Change signature of __mmap2() in direct_mmap.h to match _... by Greg Falcon <gfalcon@google.com>   - 5c02fe2031997fe064bd189fecc593395f34fae8 Change released benchmarks back to cc_test targets. by Alex Strelnikov <strel@google.com>   - 7a342ecf7a8c56276d28d94158291fe2d64b2543 Fix sanitizer example command. by Derek Mauro <dmauro@google.com>   - 0c2c55fe01cd84ab885d9f4865d7672c423680d2 Remove redundant cast; string_view::find already returns ... by Jorg Brown <jorg@google.com>
GitOrigin-RevId: 2db207ef16d548380e4f5cd1d583546d0b5f7d61

+++ absl/base/internal/direct_mmap.h
@@ -52,7 +52,7 @@
 
 // SYS_mmap and SYS_munmap are not defined in Android.
 #ifdef __BIONIC__
-extern "C" void* __mmap2(void*, size_t, int, int, int, long);
+extern "C" void* __mmap2(void*, size_t, int, int, int, size_t);
 #if defined(__NR_mmap) && !defined(SYS_mmap)
 #define SYS_mmap __NR_mmap
 #endif
